requirements.txt
asgiref==3.2.3
canvasapi==0.15.0
certifi==2019.11.28
chardet==3.0.4
Click==8.0.0
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.14
django-cors-headers==3.2.1
django-jsoneditor-widget==1.1
django-jsonfield==1.4.0
django-polymorphic==3.0.0
djangorestframework==3.11.2
gunicorn==20.0.4
idna==2.8
psycopg2-binary==2.8.6
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
sphinx==3.1.2
django-sphinx==2.2.4
django-extensions==3.0.2
ipython==7.16.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.20.5
beautifulsoup4==4.9.3
pyyaml==5.4
uritemplate==3.0.1
django-filter==2.4.0
sentry-sdk==1.5.8
black==22.10.0
